license pricing template (#1704)
* license pricing template

* add display conditions and UI flow

* add more collections and bots

* all inside the if

* more stuff

* remove delete

* seed and decimals

* re-order seeds and add a new bot after license

* like it more like this

* Fix bug we found

* add LicensePricingItem

+++ apps/platform/pkg/meta/licensetemplate.go
@@ -4,7 +4,7 @@ type LicenseTemplate struct {
 	ID           string    `json:"uesio/core.id"`
 	UniqueKey    string    `json:"uesio/core.uniquekey"`
 	App          *App      `json:"uesio/studio.app"`
-	MonthlyPrice int64     `json:"uesio/studio.monthlyprice"`
+	MonthlyPrice float64   `json:"uesio/studio.monthlyprice"`
 	AutoCreate   bool      `json:"uesio/studio.autocreate"`
 	itemMeta     *ItemMeta `json:"-"`
 	CreatedBy    *User     `json:"uesio/core.createdby"`
